fix: clang-tidy check google-runtime-int
Replace all integer types (short int long etc) with (int16_t int32_t int64_t etc).

+++ include/pika_admin.h
@@ -24,7 +24,7 @@
  */
 class SlaveofCmd : public Cmd {
  public:
-  SlaveofCmd(const std::string& name, int arity, uint16_t flag) : Cmd(name, arity, flag) {}
+  SlaveofCmd(const std::string& name, int32_t arity, uint16_t flag) : Cmd(name, arity, flag) {}
   void Do(std::shared_ptr<Slot> slot = nullptr) override;
   void Split(std::shared_ptr<Slot> slot, const HintKeys& hint_keys) override{};
   void Merge() override{};
@@ -44,7 +44,7 @@ class SlaveofCmd : public Cmd {
 
 class DbSlaveofCmd : public Cmd {
  public:
-  DbSlaveofCmd(const std::string& name, int arity, uint16_t flag) : Cmd(name, arity, flag) {}
+  DbSlaveofCmd(const std::string& name, int32_t arity, uint16_t flag) : Cmd(name, arity, flag) {}
   void Do(std::shared_ptr<Slot> slot = nullptr) override;
   void Split(std::shared_ptr<Slot> slot, const HintKeys& hint_keys) override{};
   void Merge() override{};
@@ -68,7 +68,7 @@ class DbSlaveofCmd : public Cmd {
 
 class AuthCmd : public Cmd {
  public:
-  AuthCmd(const std::string& name, int arity, uint16_t flag) : Cmd(name, arity, flag) {}
+  AuthCmd(const std::string& name, int32_t arity, uint16_t flag) : Cmd(name, arity, flag) {}
   void Do(std::shared_ptr<Slot> slot = nullptr) override;
   void Split(std::shared_ptr<Slot> slot, const HintKeys& hint_keys) override{};
   void Merge() override{};
@@ -81,7 +81,7 @@ class AuthCmd : public Cmd {
 
 class BgsaveCmd : public Cmd {
  public:
-  BgsaveCmd(const std::string& name, int arity, uint16_t flag) : Cmd(name, arity, flag) {}
+  BgsaveCmd(const std::string& name, int32_t arity, uint16_t flag) : Cmd(name, arity, flag) {}
   void Do(std::shared_ptr<Slot> slot = nullptr) override;
   void Split(std::shared_ptr<Slot> slot, const HintKeys& hint_keys) override{};
   void Merge() override{};
@@ -95,7 +95,7 @@ class BgsaveCmd : public Cmd {
 
 class CompactCmd : public Cmd {
  public:
-  CompactCmd(const std::string& name, int arity, uint16_t flag) : Cmd(name, arity, flag) {}
+  CompactCmd(const std::string& name, int32_t arity, uint16_t flag) : Cmd(name, arity, flag) {}
   void Do(std::shared_ptr<Slot> slot = nullptr) override;
   void Split(std::shared_ptr<Slot> slot, const HintKeys& hint_keys) override{};
   void Merge() override{};
@@ -113,7 +113,7 @@ class CompactCmd : public Cmd {
 
 class PurgelogstoCmd : public Cmd {
  public:
-  PurgelogstoCmd(const std::string& name, int arity, uint16_t flag) : Cmd(name, arity, flag) {}
+  PurgelogstoCmd(const std::string& name, int32_t arity, uint16_t flag) : Cmd(name, arity, flag) {}
   void Do(std::shared_ptr<Slot> slot = nullptr) override;
   void Split(std::shared_ptr<Slot> slot, const HintKeys& hint_keys) override{};
   void Merge() override{};
@@ -127,7 +127,7 @@ class PurgelogstoCmd : public Cmd {
 
 class PingCmd : public Cmd {
  public:
-  PingCmd(const std::string& name, int arity, uint16_t flag) : Cmd(name, arity, flag) {}
+  PingCmd(const std::string& name, int32_t arity, uint16_t flag) : Cmd(name, arity, flag) {}
   void Do(std::shared_ptr<Slot> slot = nullptr) override;
   void Split(std::shared_ptr<Slot> slot, const HintKeys& hint_keys) override{};
   void Merge() override{};
@@ -139,7 +139,7 @@ class PingCmd : public Cmd {
 
 class SelectCmd : public Cmd {
  public:
-  SelectCmd(const std::string& name, int arity, uint16_t flag) : Cmd(name, arity, flag) {}
+  SelectCmd(const std::string& name, int32_t arity, uint16_t flag) : Cmd(name, arity, flag) {}
   void Do(std::shared_ptr<Slot> slot = nullptr) override;
   void Split(std::shared_ptr<Slot> slot, const HintKeys& hint_keys) override{};
   void Merge() override{};
@@ -153,7 +153,7 @@ class SelectCmd : public Cmd {
 
 class FlushallCmd : public Cmd {
  public:
-  FlushallCmd(const std::string& name, int arity, uint16_t flag) : Cmd(name, arity, flag) {}
+  FlushallCmd(const std::string& name, int32_t arity, uint16_t flag) : Cmd(name, arity, flag) {}
   void Do(std::shared_ptr<Slot> slot = nullptr) override;
   void Split(std::shared_ptr<Slot> slot, const HintKeys& hint_keys) override{};
   void Merge() override{};
@@ -167,7 +167,7 @@ class FlushallCmd : public Cmd {
 
 class FlushdbCmd : public Cmd {
  public:
-  FlushdbCmd(const std::string& name, int arity, uint16_t flag) : Cmd(name, arity, flag) {}
+  FlushdbCmd(const std::string& name, int32_t arity, uint16_t flag) : Cmd(name, arity, flag) {}
   void Do(std::shared_ptr<Slot> slot = nullptr) override;
   void Split(std::shared_ptr<Slot> slot, const HintKeys& hint_keys) override{};
   void Merge() override{};
@@ -181,7 +181,7 @@ class FlushdbCmd : public Cmd {
 
 class ClientCmd : public Cmd {
  public:
-  ClientCmd(const std::string& name, int arity, uint16_t flag) : Cmd(name, arity, flag) {}
+  ClientCmd(const std::string& name, int32_t arity, uint16_t flag) : Cmd(name, arity, flag) {}
   void Do(std::shared_ptr<Slot> slot = nullptr) override;
   const static std::string CLIENT_LIST_S;
   const static std::string CLIENT_KILL_S;
@@ -214,7 +214,7 @@ class InfoCmd : public Cmd {
     kInfoCommandStats
   };
 
-  InfoCmd(const std::string& name, int arity, uint16_t flag) : Cmd(name, arity, flag) {}
+  InfoCmd(const std::string& name, int32_t arity, uint16_t flag) : Cmd(name, arity, flag) {}
   void Do(std::shared_ptr<Slot> slot = nullptr) override;
   void Split(std::shared_ptr<Slot> slot, const HintKeys& hint_keys) override{};
   void Merge() override{};
@@ -263,7 +263,7 @@ class InfoCmd : public Cmd {
 
 class ShutdownCmd : public Cmd {
  public:
-  ShutdownCmd(const std::string& name, int arity, uint16_t flag) : Cmd(name, arity, flag) {}
+  ShutdownCmd(const std::string& name, int32_t arity, uint16_t flag) : Cmd(name, arity, flag) {}
   void Do(std::shared_ptr<Slot> slot = nullptr) override;
   void Split(std::shared_ptr<Slot> slot, const HintKeys& hint_keys) override{};
   void Merge() override{};
@@ -275,7 +275,7 @@ class ShutdownCmd : public Cmd {
 
 class ConfigCmd : public Cmd {
  public:
-  ConfigCmd(const std::string& name, int arity, uint16_t flag) : Cmd(name, arity, flag) {}
+  ConfigCmd(const std::string& name, int32_t arity, uint16_t flag) : Cmd(name, arity, flag) {}
   void Do(std::shared_ptr<Slot> slot = nullptr) override;
   void Split(std::shared_ptr<Slot> slot, const HintKeys& hint_keys) override{};
   void Merge() override{};
@@ -292,7 +292,7 @@ class ConfigCmd : public Cmd {
 
 class MonitorCmd : public Cmd {
  public:
-  MonitorCmd(const std::string& name, int arity, uint16_t flag) : Cmd(name, arity, flag) {}
+  MonitorCmd(const std::string& name, int32_t arity, uint16_t flag) : Cmd(name, arity, flag) {}
   void Do(std::shared_ptr<Slot> slot = nullptr) override;
   void Split(std::shared_ptr<Slot> slot, const HintKeys& hint_keys) override{};
   void Merge() override{};
@@ -304,7 +304,7 @@ class MonitorCmd : public Cmd {
 
 class DbsizeCmd : public Cmd {
  public:
-  DbsizeCmd(const std::string& name, int arity, uint16_t flag) : Cmd(name, arity, flag) {}
+  DbsizeCmd(const std::string& name, int32_t arity, uint16_t flag) : Cmd(name, arity, flag) {}
   void Do(std::shared_ptr<Slot> slot = nullptr) override;
   void Split(std::shared_ptr<Slot> slot, const HintKeys& hint_keys) override{};
   void Merge() override{};
@@ -316,7 +316,7 @@ class DbsizeCmd : public Cmd {
 
 class TimeCmd : public Cmd {
  public:
-  TimeCmd(const std::string& name, int arity, uint16_t flag) : Cmd(name, arity, flag) {}
+  TimeCmd(const std::string& name, int32_t arity, uint16_t flag) : Cmd(name, arity, flag) {}
   void Do(std::shared_ptr<Slot> slot = nullptr) override;
   void Split(std::shared_ptr<Slot> slot, const HintKeys& hint_keys) override{};
   void Merge() override{};
@@ -328,7 +328,7 @@ class TimeCmd : public Cmd {
 
 class DelbackupCmd : public Cmd {
  public:
-  DelbackupCmd(const std::string& name, int arity, uint16_t flag) : Cmd(name, arity, flag) {}
+  DelbackupCmd(const std::string& name, int32_t arity, uint16_t flag) : Cmd(name, arity, flag) {}
   void Do(std::shared_ptr<Slot> slot = nullptr) override;
   void Split(std::shared_ptr<Slot> slot, const HintKeys& hint_keys) override{};
   void Merge() override{};
@@ -340,7 +340,7 @@ class DelbackupCmd : public Cmd {
 
 class EchoCmd : public Cmd {
  public:
-  EchoCmd(const std::string& name, int arity, uint16_t flag) : Cmd(name, arity, flag) {}
+  EchoCmd(const std::string& name, int32_t arity, uint16_t flag) : Cmd(name, arity, flag) {}
   void Merge() override{};
   void Do(std::shared_ptr<Slot> slot = nullptr) override;
   void Split(std::shared_ptr<Slot> slot, const HintKeys& hint_keys) override{};
@@ -353,7 +353,7 @@ class EchoCmd : public Cmd {
 
 class ScandbCmd : public Cmd {
  public:
-  ScandbCmd(const std::string& name, int arity, uint16_t flag) : Cmd(name, arity, flag) {}
+  ScandbCmd(const std::string& name, int32_t arity, uint16_t flag) : Cmd(name, arity, flag) {}
   void Do(std::shared_ptr<Slot> slot = nullptr) override;
   void Split(std::shared_ptr<Slot> slot, const HintKeys& hint_keys) override{};
   void Merge() override{};
@@ -368,7 +368,7 @@ class ScandbCmd : public Cmd {
 class SlowlogCmd : public Cmd {
  public:
   enum SlowlogCondition { kGET, kLEN, kRESET };
-  SlowlogCmd(const std::string& name, int arity, uint16_t flag) : Cmd(name, arity, flag) {}
+  SlowlogCmd(const std::string& name, int32_t arity, uint16_t flag) : Cmd(name, arity, flag) {}
   void Do(std::shared_ptr<Slot> slot = nullptr) override;
   void Split(std::shared_ptr<Slot> slot, const HintKeys& hint_keys) override{};
   void Merge() override{};
@@ -386,7 +386,7 @@ class SlowlogCmd : public Cmd {
 
 class PaddingCmd : public Cmd {
  public:
-  PaddingCmd(const std::string& name, int arity, uint16_t flag) : Cmd(name, arity, flag) {}
+  PaddingCmd(const std::string& name, int32_t arity, uint16_t flag) : Cmd(name, arity, flag) {}
   void Do(std::shared_ptr<Slot> slot = nullptr) override;
   void Split(std::shared_ptr<Slot> slot, const HintKeys& hint_keys) override{};
   void Merge() override{};
@@ -400,7 +400,7 @@ class PaddingCmd : public Cmd {
 
 class PKPatternMatchDelCmd : public Cmd {
  public:
-  PKPatternMatchDelCmd(const std::string& name, int arity, uint16_t flag) : Cmd(name, arity, flag) {}
+  PKPatternMatchDelCmd(const std::string& name, int32_t arity, uint16_t flag) : Cmd(name, arity, flag) {}
   void Do(std::shared_ptr<Slot> slot = nullptr) override;
   void Split(std::shared_ptr<Slot> slot, const HintKeys& hint_keys) override{};
   void Merge() override{};
@@ -415,7 +415,7 @@ class PKPatternMatchDelCmd : public Cmd {
 class DummyCmd : public Cmd {
  public:
   DummyCmd() : Cmd("", 0, 0) {}
-  DummyCmd(const std::string& name, int arity, uint16_t flag) : Cmd(name, arity, flag) {}
+  DummyCmd(const std::string& name, int32_t arity, uint16_t flag) : Cmd(name, arity, flag) {}
   void Do(std::shared_ptr<Slot> slot = nullptr) override;
   void Split(std::shared_ptr<Slot> slot, const HintKeys& hint_keys) override{};
   void Merge() override{};
@@ -427,7 +427,7 @@ class DummyCmd : public Cmd {
 
 class QuitCmd : public Cmd {
  public:
-  QuitCmd(const std::string& name, int arity, uint16_t flag) : Cmd(name, arity, flag) {}
+  QuitCmd(const std::string& name, int32_t arity, uint16_t flag) : Cmd(name, arity, flag) {}
   void Do(std::shared_ptr<Slot> slot = nullptr) override;
   void Split(std::shared_ptr<Slot> slot, const HintKeys& hint_keys) override{};
   void Merge() override{};
@@ -439,7 +439,7 @@ class QuitCmd : public Cmd {
 
 class HelloCmd : public Cmd {
  public:
-  HelloCmd(const std::string& name, int arity, uint16_t flag) : Cmd(name, arity, flag) {}
+  HelloCmd(const std::string& name, int32_t arity, uint16_t flag) : Cmd(name, arity, flag) {}
   void Do(std::shared_ptr<Slot> slot = nullptr) override;
   void Split(std::shared_ptr<Slot> slot, const HintKeys& hint_keys) override{};
   void Merge() override{};
@@ -452,7 +452,7 @@ class HelloCmd : public Cmd {
 #ifdef WITH_COMMAND_DOCS
 class CommandCmd : public Cmd {
  public:
-  CommandCmd(const std::string& name, int arity, uint16_t flag) : Cmd(name, arity, flag) {}
+  CommandCmd(const std::string& name, int32_t arity, uint16_t flag) : Cmd(name, arity, flag) {}
   void Do(std::shared_ptr<Slot> slot = nullptr) override;
   void Split(std::shared_ptr<Slot> slot, const HintKeys& hint_keys) override{};
   void Merge() override{};
@@ -464,7 +464,7 @@ class CommandCmd : public Cmd {
     bool operator()(const std::string&, const std::string&) const;
 
    private:
-    const static std::unordered_map<std::string, int> kFieldNameOrder;
+    const static std::unordered_map<std::string, int32_t> kFieldNameOrder;
   };
 
   class Encodable;
@@ -482,15 +482,15 @@ class CommandCmd : public Cmd {
 
   class EncodableInt : public Encodable {
    public:
-    EncodableInt(int value) : value_(value) {}
-    EncodableInt(unsigned long long value) : value_(value) {}
+    EncodableInt(int32_t value) : value_(value) {}
+    EncodableInt(uint64_t value) : value_(value) {}
 
    protected:
     CmdRes& EncodeTo(CmdRes& res) const override;
     EncodablePtr MergeFrom(const EncodablePtr& other) const override;
 
    private:
-    int value_;
+    int32_t value_;
   };
 
   class EncodableString : public Encodable {
@@ -585,7 +585,7 @@ class CommandCmd : public Cmd {
   const static std::unordered_map<std::string, EncodablePtr> kCommandDocs;
 };
 
-static CommandCmd::EncodablePtr operator""_RedisInt(unsigned long long value);
+static CommandCmd::EncodablePtr operator""_RedisInt(uint64_t value);
 static CommandCmd::EncodablePtr operator""_RedisString(const char* value);
 static CommandCmd::EncodablePtr operator""_RedisStatus(const char* value);
 static CommandCmd::EncodablePtr RedisMap(CommandCmd::EncodableMap::RedisMap values);
