+++ src/net/include/redis_conn.h
@@ -23,13 +23,13 @@ enum HandleType { kSynchronous, kAsynchronous };
 
 class RedisConn : public NetConn {
  public:
-  RedisConn(int fd, const std::string& ip_port, Thread* thread, NetMultiplexer* net_mpx = nullptr,
-            const HandleType& handle_type = kSynchronous, int rbuf_max_len = REDIS_MAX_MESSAGE);
+  RedisConn(int32_t fd, const std::string& ip_port, Thread* thread, NetMultiplexer* net_mpx = nullptr,
+            const HandleType& handle_type = kSynchronous, int32_t rbuf_max_len = REDIS_MAX_MESSAGE);
   ~RedisConn() override;
 
   ReadStatus GetRequest() override;
   WriteStatus SendReply() override;
-  int WriteResp(const std::string& resp) override;
+  int32_t WriteResp(const std::string& resp) override;
 
   void TryResizeBuffer() override;
   void SetHandleType(const HandleType& handle_type);
@@ -38,29 +38,29 @@ class RedisConn : public NetConn {
   virtual void ProcessRedisCmds(const std::vector<RedisCmdArgsType>& argvs, bool async, std::string* response);
   void NotifyEpoll(bool success);
 
-  virtual int DealMessage(const RedisCmdArgsType& argv, std::string* response) = 0;
+  virtual int32_t DealMessage(const RedisCmdArgsType& argv, std::string* response) = 0;
   virtual const std::string& GetCurrentTable() = 0;
 
  private:
-  static int ParserDealMessageCb(RedisParser* parser, const RedisCmdArgsType& argv);
-  static int ParserCompleteCb(RedisParser* parser, const std::vector<RedisCmdArgsType>& argvs);
+  static int32_t ParserDealMessageCb(RedisParser* parser, const RedisCmdArgsType& argv);
+  static int32_t ParserCompleteCb(RedisParser* parser, const std::vector<RedisCmdArgsType>& argvs);
   ReadStatus ParseRedisParserStatus(RedisParserStatus status);
 
   HandleType handle_type_ = kSynchronous;
 
   char* rbuf_ = nullptr;
-  int rbuf_len_ = 0;
-  int rbuf_max_len_ = 0;
-  int msg_peak_ = 0;
-  int command_len_ = 0;
+  int32_t rbuf_len_ = 0;
+  int32_t rbuf_max_len_ = 0;
+  int32_t msg_peak_ = 0;
+  int32_t command_len_ = 0;
 
   uint32_t wbuf_pos_ = 0;
   std::string response_;
 
   // For Redis Protocol parser
-  int last_read_pos_ = -1;
+  int32_t last_read_pos_ = -1;
   RedisParser redis_parser_;
-  long bulk_len_ = -1;
+  int32_t bulk_len_ = -1;
 };
 
 }  // namespace net
